+++ lokalise/models/v1/audit_log.py
@@ -1,0 +1,35 @@
+"""
+lokalise.models.audit_log
+~~~~~~~~~~~~~~~~~~~~~~~~~
+Module containing audit log.
+"""
+
+from typing import ClassVar
+
+from ..base_model import BaseModel
+
+
+class AuditLogModel(BaseModel):
+    """Describes audit log model."""
+
+    ATTRS: ClassVar[list[str]] = [
+        "class_uid",
+        "class_name",
+        "category_uid",
+        "category_name",
+        "activity_id",
+        "activity_name",
+        "type_uid",
+        "type_name",
+        "severity_id",
+        "severity",
+        "status_id",
+        "status",
+        "time",
+        "metadata",
+        "actor",
+        "src_endpoint",
+        "http_request",
+        "enrichments",
+        "unmapped",
+    ]